Add files via upload
This schedule details all of the tasks that I had to manage and assign roles to within the programming department. Since I was the lead on this project I had to manage my own department as well as keep in close contact with the game designer and other department leads.